pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length = 100

[tool.isort]
known_django = ["django"]
line_length = 100
profile = "black"
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"

[tool.poetry]
authors = ["DeMiA inc."]
description = ""
name = "event_line_bot_numeron"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "^4.0.3"
dj-database-url = "^0.5.0"
django-sass-processor = {extras = ["management-command"], version = "^1.1"}
djangorestframework = "^3.13.1"
libsass = "^0.21.0"
python = "^3.8"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.17.0"
toml-sort = "^0.19.0"
